+++ backend/src/main/java/com/woowacourse/gongseek/article/domain/repository/dto/ArticlePreviewDto.java
@@ -2,8 +2,8 @@ package com.woowacourse.gongseek.article.domain.repository.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.woowacourse.gongseek.article.domain.Category;
+import com.woowacourse.gongseek.member.application.dto.AuthorDto;
 import com.woowacourse.gongseek.member.domain.Member;
-import com.woowacourse.gongseek.member.presentation.dto.AuthorDto;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AccessLevel;
