+++ backend/src/main/java/com/woowacourse/gongseek/comment/application/dto/CommentResponse.java
@@ -2,7 +2,7 @@ package com.woowacourse.gongseek.comment.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.woowacourse.gongseek.comment.domain.Comment;
-import com.woowacourse.gongseek.member.presentation.dto.AuthorDto;
+import com.woowacourse.gongseek.member.application.dto.AuthorDto;
 import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
