+++ backend/src/main/java/com/woowacourse/gongseek/member/application/MemberService.java
@@ -5,14 +5,14 @@ import com.woowacourse.gongseek.article.domain.repository.dto.MyPagePreviewDto;
 import com.woowacourse.gongseek.auth.application.dto.AppMember;
 import com.woowacourse.gongseek.comment.domain.Comment;
 import com.woowacourse.gongseek.comment.domain.repository.CommentRepository;
+import com.woowacourse.gongseek.member.application.dto.MemberDto;
+import com.woowacourse.gongseek.member.application.dto.MemberUpdateRequest;
+import com.woowacourse.gongseek.member.application.dto.MemberUpdateResponse;
+import com.woowacourse.gongseek.member.application.dto.MyPageArticlesResponse;
+import com.woowacourse.gongseek.member.application.dto.MyPageCommentsResponse;
 import com.woowacourse.gongseek.member.domain.Member;
 import com.woowacourse.gongseek.member.domain.repository.MemberRepository;
 import com.woowacourse.gongseek.member.exception.MemberNotFoundException;
-import com.woowacourse.gongseek.member.presentation.dto.MemberDto;
-import com.woowacourse.gongseek.member.presentation.dto.MemberUpdateRequest;
-import com.woowacourse.gongseek.member.presentation.dto.MemberUpdateResponse;
-import com.woowacourse.gongseek.member.presentation.dto.MyPageArticlesResponse;
-import com.woowacourse.gongseek.member.presentation.dto.MyPageCommentsResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
