+++ backend/src/main/java/com/woowacourse/gongseek/member/application/dto/MemberUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.gongseek.member.presentation.dto;
+package com.woowacourse.gongseek.member.application.dto;
 
 import javax.validation.constraints.NotNull;
 import lombok.AccessLevel;
