+++ backend/src/test/java/com/woowacourse/gongseek/comment/application/CommentServiceTest.java
@@ -18,10 +18,10 @@ import com.woowacourse.gongseek.comment.application.dto.CommentUpdateRequest;
 import com.woowacourse.gongseek.comment.domain.Comment;
 import com.woowacourse.gongseek.comment.domain.repository.CommentRepository;
 import com.woowacourse.gongseek.comment.exception.CommentNotFoundException;
+import com.woowacourse.gongseek.member.application.dto.MemberDto;
 import com.woowacourse.gongseek.member.domain.Member;
 import com.woowacourse.gongseek.member.domain.repository.MemberRepository;
 import com.woowacourse.gongseek.member.exception.MemberNotFoundException;
-import com.woowacourse.gongseek.member.presentation.dto.MemberDto;
 import com.woowacourse.gongseek.support.IntegrationTest;
 import java.util.List;
 import java.util.stream.Collectors;
